Rename `rustc_codegen_gcc/errors.rs` into `rustc_codegen_gcc/diagnostics.rs`
