+++ src/back/write.rs
@@ -12,7 +12,7 @@ use rustc_session::config::OutputType;
 use rustc_target::spec::SplitDebuginfo;
 
 use crate::base::add_pic_option;
-use crate::errors::CopyBitcode;
+use crate::diagnostics::CopyBitcode;
 use crate::{GccContext, LtoMode};
 
 pub(crate) fn codegen(
